MO-752 Add hard-coded movementTime field to emulator
